refactor: 로또 서비스 로직을 비즈니스 로직과 도메인 로직으로 분리 후 수정
